restructure priv risk score to be mia on point basis #151 (#180)
* restructure priv risk score to be mia on point basis #151

* grrrr forgot to git add any files ._.

* Small fixes. #151

* Small fixes. #151

* small fixes #154

* changed notebooks to work with MIA on point basis class structure

* fixed mia on point basis dataset sample size

* Use only first 100 data points

* final changes the notebooks
