+++ privacy_evaluator/metrics/__init__.py
@@ -2,4 +2,3 @@
 Module providing metrics.
 """
 from .basics import accuracy
-from .privacy_risk_score import compute_privacy_risk_score
